+++ apps/api-journeys/src/modules/block/block.spec.ts
@@ -39,107 +39,6 @@ describe('BlockModule', () => {
     })
   }
 
-  it('returns ButtonBlock', async () => {
-    const parentBlockId = uuidv4()
-    const button: Block = {
-      id: uuidv4(),
-      journeyId,
-      blockType: 'ButtonBlock',
-      parentBlockId,
-      parentOrder: 1,
-      extraAttrs: {
-        label: 'label',
-        variant: 'contained',
-        color: 'primary',
-        size: 'large',
-        startIcon: {
-          name: 'ArrowForward',
-          color: 'secondary',
-          size: 'lg'
-        },
-        endIcon: {
-          name: 'LockOpen',
-          color: 'action',
-          size: 'xl'
-        },
-        action: {
-          gtmEventName: 'gtmEventName',
-          url: 'https://jesusfilm.org',
-          target: 'target'
-        }
-      }
-    }
-    dbMock.block.findMany.mockResolvedValue([button])
-    const { data } = await query(gql`
-      query ($id: ID!) {
-        journey(id: $id) {
-          blocks {
-            id
-            __typename
-            parentBlockId
-            ... on ButtonBlock {
-              label
-              variant
-              color
-              size
-              startIcon {
-                name
-                color
-                size
-              }
-              endIcon {
-                name
-                color
-                size
-              }
-              action {
-                __typename
-                gtmEventName
-                ... on NavigateToBlockAction {
-                  blockId
-                }
-                ... on NavigateToJourneyAction {
-                  journeyId
-                }
-                ... on LinkAction {
-                  url
-                  target
-                }
-              }
-            }
-          }
-        }
-      }
-    `)
-    expect(data?.journey.blocks).toEqual([
-      {
-        id: button.id,
-        __typename: 'ButtonBlock',
-        parentBlockId,
-        label: 'label',
-        variant: 'contained',
-        color: 'primary',
-        size: 'large',
-        startIcon: {
-          name: 'ArrowForward',
-          color: 'secondary',
-          size: 'lg'
-        },
-        endIcon: {
-          name: 'LockOpen',
-          color: 'action',
-          size: 'xl'
-        },
-        action: {
-          __typename: 'LinkAction',
-          gtmEventName: 'gtmEventName',
-          url: 'https://jesusfilm.org',
-          target: 'target'
-        }
-      }
-    ])
-  })
-
   it('returns CardBlock', async () => {
     const parentBlockId = uuidv4()
     const coverBlockId = uuidv4()
@@ -270,52 +169,6 @@ describe('BlockModule', () => {
         parentBlockId,
         locked: true,
         nextBlockId
-      }
-    ])
-  })
-
-  it('returns TypographyBlock', async () => {
-    const parentBlockId = uuidv4()
-    const typography: Block = {
-      id: uuidv4(),
-      journeyId,
-      blockType: 'TypographyBlock',
-      parentBlockId,
-      parentOrder: 7,
-      extraAttrs: {
-        content: 'text',
-        variant: 'h2',
-        color: 'primary',
-        align: 'left'
-      }
-    }
-    dbMock.block.findMany.mockResolvedValue([typography])
-    const { data } = await query(gql`
-      query ($id: ID!) {
-        journey(id: $id) {
-          blocks {
-            id
-            __typename
-            parentBlockId
-            ... on TypographyBlock {
-              content
-              variant
-              color
-              align
-            }
-          }
-        }
-      }
-    `)
-    expect(data?.journey.blocks).toEqual([
-      {
-        id: typography.id,
-        __typename: 'TypographyBlock',
-        parentBlockId,
-        content: 'text',
-        variant: 'h2',
-        color: 'primary',
-        align: 'left'
       }
     ])
   })
