+++ apps/api-journeys/src/modules/icon/index.ts
@@ -1,0 +1,1 @@
+export { iconModule } from './icon'
