+++ apps/api-journeys/src/modules/index.ts
@@ -1,7 +1,10 @@
 export { actionModule } from './action'
 export { blockModule } from './block'
+export { buttonModule } from './button'
+export { iconModule } from './icon'
 export { journeyModule } from './journey'
 export { radioQuestionModule } from './radioQuestion'
 export { responseModule } from './response'
 export { signUpModule } from './signUp'
+export { typographyModule } from './typography'
 export { videoModule } from './video'
